refactor(clients): extract helpers + BentoStatsClients (Phase 3.5 audit)
Phase 3.5 brainstorm Égypte 2026-05 — refacto #3/3 ClientsPage.
ClientsPage.tsx : 1425 → 1080 lignes (-24%). Zéro changement de comportement.

NOUVEAU
- src/components/clients/clientsListHelpers.ts (145 L) : helpers de
  présentation (getOwnerAvatarColors, getClientStatusInfo, isOverdue,
  getDaysUntil, getRelativeTime, getLastAssessmentTime, exportClientsCsv)
  + LIFECYCLE_TONE_TO_COLORS. Fonctions pures, zéro dépendance React.
- src/components/clients/BentoStatsClients.tsx (257 L) : 3 cards stats
  (visibles / responsables / relances) avec animations CSS. Props : 3
  numéros (visibleCount, ownersCount, relanceCount).

MIGRE
- ClientsPage.tsx : imports remontés (clientsListHelpers + BentoStatsClients
  + getInitials shared util), bento stats section (235 L) remplacée par
  <BentoStatsClients />, anciens helpers locaux supprimés (-135 L), import
  type User/Client/LIFECYCLE_LABELS retiré (devenu inutile).

NON FAIT (raison : audit ciblait sub-800 mais risque > gain pour ce
livrable)
- Extraction du body 404→1070 (clients tab) en <ClientsListView /> à part
  entière nécessiterait de bouger 30+ state/effects/computed/handlers,
  scope plus large + risque régression élevé. À planifier en chantier
  dédié séparé si Thomas valide ROI.

Build prod : npm run build ✅ (chunk ClientsPage 78.19 kB, identique avant).
